remarks and active job extraction to feature
