+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Internal refactoring of routing validations flow.
 - **[Breaking]** Disable the root `manual_offset_management` setting and require it to be configured per topic. This is part of "topic features" configuration extraction for better code organization.
 - Reorganize how routing related features are represented internally to simplify features management.
+- Fix a case where routing tags would not be injected when given routing definition would not be used with a block
 
 ## 2.0.15 (2022-10-20)
 - Sanitize admin config prior to any admin action.
