+++ spec/lib/karafka/routing/features/active_job_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RSpec.describe_current do
+  it { expect(described_class).to be < ::Karafka::Routing::Features::Base }
+end
