+++ spec/lib/karafka/routing/features/base/expander_spec.rb
